+++ bridge/contracts/SideToken.sol
@@ -18,7 +18,7 @@ contract SideToken is DetailedERC20, StandardToken, Transferable {
         manager = _manager;
     }
     
-    function acceptTransfer(address receiver, uint amount) public onlyManager returns(bool) {
+    function acceptTransfer(address receiver, uint256 amount) public onlyManager returns(bool) {
         totalSupply_ += amount;
         balances[receiver] += amount;
